Organise documentation into modules.
