+++ README.rst
@@ -12,20 +12,7 @@
 
 A collection of functionality useful for scientific computing.
 
-Interface
----------
+.. toctree::
 
-.. automodule:: sciutils
-   :members:
-
-Statistical distributions and utility functions
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-.. automodule:: sciutils.stats
-   :members:
-
-Utility functions for plotting
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-.. automodule:: sciutils.plot
-   :members:
+   docs/plot
+   docs/stats
