+++ docs/plot.rst
@@ -1,0 +1,5 @@
+Utility functions for plotting
+==============================
+
+.. automodule:: sciutils.plot
+   :members:
