+++ docs/stats.rst
@@ -1,0 +1,5 @@
+Statistical distributions and utility functions
+===============================================
+
+.. automodule:: sciutils.stats
+   :members:
